refactor kernel detector
