+++ nn_meter/kernel_detector/kernel_detector.py
@@ -3,8 +3,8 @@
 from nn_meter.utils.graph_tool import ModelGraph
 from .utils.constants import DUMMY_TYPES
 from .utils.ir_tools import convert_nodes
-from .rulelib.rule_reader import RuleReader
-from .rulelib.rule_splitter import RuleSplitter
+from .rule_reader import RuleReader
+from .rule_splitter import RuleSplitter
 
 
 class KernelDetector:
