+++ nn_meter/kernel_detector/rule_reader.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 import json
-from ..fusionlib import get_fusion_unit
+from .fusionlib import get_fusion_unit
 from nn_meter.utils.graph_tool import ModelGraph
